Ran compiler

+++ 0x12-singly_linked_lists/1-main.c
@@ -1,6 +1,4 @@
-#include <stdlib.h>
 #include <string.h>
-#include <stdio.h>
 #include "lists.h"
 
 /**
@@ -27,7 +25,7 @@ int main(void)
     new->next = head;
     head = new;
     n = list_len(head);
-    printf("-> %lu elements\n", n);
+    printf("-> %lu elements\n", (long unsigned int)n);
     free(new->str);
     free(new);
     return (0);
